pyproject.toml
[project]
name = "mcp-server-tiny-print"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "bleak>=0.21.0",
    "pillow>=10.0.0",
    "mcp[cli]>=1.0.0",
    "markdown>=3.4.0",
]

[project.scripts]
mcp-server-tiny-print = "mcp_tiny_print.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_tiny_print"]
